+++ src/components/ui/sidebar/app-sidebar.tsx
@@ -1,8 +1,6 @@
 import { DashboardIcon } from "@radix-ui/react-icons";
-import { TFunction } from "i18next";
 import { Link, usePathParams } from "raviger";
 import * as React from "react";
-import { useTranslation } from "react-i18next";
 
 import {
   Sidebar,
@@ -14,121 +12,32 @@ import {
   SidebarMenuItem,
   SidebarRail,
 } from "@/components/ui/sidebar";
+import { FacilityNav } from "@/components/ui/sidebar/facility-nav";
 import { FacilitySwitcher } from "@/components/ui/sidebar/facility-switcher";
-import { NavMain } from "@/components/ui/sidebar/nav-main";
 import {
   FacilityNavUser,
   PatientNavUser,
 } from "@/components/ui/sidebar/nav-user";
+import { OrgNav } from "@/components/ui/sidebar/org-nav";
 import { OrganizationSwitcher } from "@/components/ui/sidebar/organization-switcher";
+import { PatientNav } from "@/components/ui/sidebar/patient-nav";
 
 import { UserFacilityModel, UserModel } from "@/components/Users/models";
-
-import { PatientUserContextType } from "@/Providers/PatientUserProvider";
-import { AppointmentPatient } from "@/pages/Patient/Utils";
-import { Organization } from "@/types/organization/organization";
-
-import { PatientSwitcher } from "./patient-switcher";
-
-interface NavigationLink {
-  name: string;
-  url: string;
-  icon?: string;
-}
 
 interface AppSidebarProps extends React.ComponentProps<typeof Sidebar> {
   user?: UserModel;
   facilitySidebar?: boolean;
-  patientUserContext?: PatientUserContextType;
+  sidebarFor?: SidebarFor;
 }
 
-function generateFacilityLinks(
-  selectedFacility: UserFacilityModel | null,
-  t: TFunction,
-  // TODO: switch to UserBase once getcurrentuser serializer is updated
-  user?: UserModel,
-) {
-  if (!selectedFacility) return [];
-
-  const baseUrl = `/facility/${selectedFacility.id}`;
-  const links: NavigationLink[] = [
-    { name: t("facility"), url: baseUrl, icon: "d-hospital" },
-    {
-      name: t("appointments"),
-      url: `${baseUrl}/appointments`,
-      icon: "d-calendar",
-    },
-    {
-      name: t("search_patients"),
-      url: `${baseUrl}/patients`,
-      icon: "d-patient",
-    },
-    { name: t("encounters"), url: `${baseUrl}/encounters`, icon: "d-patient" },
-    // { name: t("assets"), url: `${baseUrl}/assets`, icon: "d-folder" },
-    { name: t("resource"), url: "/resource", icon: "d-book-open" },
-    { name: t("users"), url: `${baseUrl}/users`, icon: "d-people" },
-    {
-      name: t("organization"),
-      url: `${baseUrl}/organization`,
-      icon: "d-book-open",
-    },
-  ];
-
-  if (user) {
-    links.push({
-      name: t("schedules"),
-      url: `${baseUrl}/users/${user.username}/availability`,
-      icon: "d-calendar",
-    });
-  }
-
-  return links;
-}
-
-function generateOrganizationLinks(
-  organizations: Organization[],
-): NavigationLink[] {
-  return organizations.map((org) => ({
-    name: org.name,
-    url: `/organization/${org.id}`,
-  }));
-}
-
-function generatePatientLinks(
-  selectedUser: AppointmentPatient | null,
-  t: TFunction,
-): NavigationLink[] {
-  if (!selectedUser) return [];
-
-  const { geo_organization } = selectedUser;
-  let parentOrganization = geo_organization?.parent;
-  while (parentOrganization?.parent) {
-    if (parentOrganization.level_cache === 1) {
-      break;
-    }
-    parentOrganization = parentOrganization.parent;
-  }
-
-  const queryParams = new URLSearchParams();
-
-  if (parentOrganization) {
-    queryParams.set("organization", String(parentOrganization?.id));
-  }
-
-  return [
-    { name: t("appointments"), url: "/patient/home", icon: "d-patient" },
-    {
-      name: t("nearby_facilities"),
-      url: `/nearby_facilities/?${queryParams.toString()}`,
-      icon: "d-patient",
-    },
-  ];
+export enum SidebarFor {
+  FACILITY = "facility",
+  PATIENT = "patient",
 }
 
 export function AppSidebar({
   user,
-  patientUserContext,
-  facilitySidebar = true,
+  sidebarFor = SidebarFor.FACILITY,
   ...props
 }: AppSidebarProps) {
   const exactMatch = usePathParams("/facility/:facilityId");
@@ -139,10 +48,11 @@ export function AppSidebar({
   const orgSubpathMatch = usePathParams("/organization/:id/*");
   const organizationId = orgMatch?.id || orgSubpathMatch?.id;
 
+  const facilitySidebar = sidebarFor === SidebarFor.FACILITY;
+  const patientSidebar = sidebarFor === SidebarFor.PATIENT;
+
   const [selectedFacility, setSelectedFacility] =
     React.useState<UserFacilityModel | null>(null);
-
-  const { t } = useTranslation();
 
   const selectedOrganization = React.useMemo(() => {
     if (!user?.organizations || !organizationId) return undefined;
@@ -207,34 +117,17 @@ export function AppSidebar({
 
       <SidebarContent>
         {facilitySidebar && !selectedOrganization && (
-          <NavMain links={generateFacilityLinks(selectedFacility, t, user)} />
+          <FacilityNav selectedFacility={selectedFacility} user={user} />
         )}
         {selectedOrganization && (
-          <NavMain
-            links={generateOrganizationLinks(user?.organizations || [])}
-          />
+          <OrgNav organizations={user?.organizations || []} />
         )}
-        {patientUserContext && (
-          <>
-            <PatientSwitcher />
-            <NavMain
-              links={generatePatientLinks(
-                patientUserContext.selectedPatient,
-                t,
-              )}
-            />
-          </>
-        )}
+        {patientSidebar && <PatientNav />}
       </SidebarContent>
 
       <SidebarFooter>
         {(facilitySidebar || selectedOrganization) && <FacilityNavUser />}
-        {patientUserContext && (
-          <PatientNavUser
-            patient={patientUserContext.selectedPatient}
-            phoneNumber={patientUserContext.tokenData.phoneNumber}
-          />
-        )}
+        {patientSidebar && <PatientNavUser />}
       </SidebarFooter>
 
       <SidebarRail />
