+++ src/components/ui/sidebar/nav-user.tsx
@@ -25,8 +25,7 @@ import { Avatar } from "@/components/Common/Avatar";
 
 import useAuthUser, { useAuthContext } from "@/hooks/useAuthUser";
 import { usePatientSignOut } from "@/hooks/usePatientSignOut";
-
-import { AppointmentPatient } from "@/pages/Patient/Utils";
+import { usePatientContext } from "@/hooks/usePatientUser";
 
 export function FacilityNavUser() {
   const { t } = useTranslation();
@@ -117,16 +116,14 @@ export function FacilityNavUser() {
   );
 }
 
-export function PatientNavUser({
-  patient,
-  phoneNumber,
-}: {
-  patient: AppointmentPatient | null;
-  phoneNumber: string;
-}) {
+export function PatientNavUser() {
   const { t } = useTranslation();
   const { isMobile, open } = useSidebar();
   const signOut = usePatientSignOut();
+  const patientUserContext = usePatientContext();
+
+  const patient = patientUserContext?.selectedPatient;
+  const phoneNumber = patientUserContext?.tokenData.phoneNumber;
 
   return (
     <SidebarMenu>
